chore: remove the Telegram integration
Removes the Telegram cloud-connect feature end to end: the bot API client,
update dispatcher, webhook and org routes, the pairing/chat store, the
dashboard dialog and quick-add tile, and the connection hooks.

The cloud-worker plumbing that lived in telegram-worker.ts was never
Telegram-specific and is still needed by the automations cloud agent
executor, so it moves to workers/worker-access.ts as loadCloudWorkerAccess.
The prompt/session/snapshot machinery had no remaining consumer and is
dropped with the rest.

The il_* tables, migration 0033 and the tg* typeid prefixes are handled in
a separate commit. Changelog entries and translated READMEs are left as
written: they record what shipped at the time.
