+++ ee/apps/den-api/.env.example
@@ -16,7 +16,7 @@ DEN_DB_ENCRYPTION_KEY=
 BETTER_AUTH_URL=http://localhost:3005
 DEN_BETTER_AUTH_TRUSTED_ORIGINS=http://localhost:3005,http://127.0.0.1:3005
 # Public API base; an HTTPS reverse-proxy prefix such as /api/den is supported.
-# OAuth callbacks and /v1/webhooks/telegram must bypass upstream SSO/Access challenges.
+# OAuth callbacks must bypass upstream SSO/Access challenges.
 DEN_API_PUBLIC_URL=http://localhost:8790
 # Optional protocol-compatible endpoint overrides for connector tests. OAuth
 # overrides must contain a literal {tenantId} path placeholder. Leave blank in
@@ -28,9 +28,6 @@ DEN_MICROSOFT_GRAPH_BASE_URL=
 # the base URL override is intended for compatible self-hosted/test endpoints.
 DEN_DESKTOP_RELEASES_MODE=github
 DEN_DESKTOP_RELEASES_BASE_URL=
-# Optional protocol-compatible Telegram Bot API root for local/Daytona tests.
-# Leave blank in production to use https://api.telegram.org.
-DEN_TELEGRAM_API_ROOT=
 DEN_DESKTOP_DEN_BASE_URL=http://localhost:3005/api/den
 DEN_MARKETING_URL=http://localhost:3005
 DEN_MCP_RESOURCE_URL=http://localhost:8790/mcp
