+++ ee/apps/den-api/src/routes/org/delete-organization.ts
@@ -58,10 +58,6 @@ import {
   SsoProviderTable,
   TeamMemberTable,
   TeamTable,
-  TelegramChatBindingTable,
-  TelegramConnectionTable,
-  TelegramPairingTable,
-  TelegramUpdateTable,
   TelemetryEventTable,
   TelemetrySessionDimensionTable,
   WorkerBundleTable,
@@ -441,17 +437,6 @@ export function registerDeleteOrganizationRoutes<T extends { Variables: OrgRoute
           await tx.delete(InferenceUsageLedgerBucketChargeTable).where(inArray(InferenceUsageLedgerBucketChargeTable.ledger_entry_id, ledgerEntryIds))
         }
 
-        const telegramConnectionIds = (await tx
-          .select({ id: TelegramConnectionTable.id })
-          .from(TelegramConnectionTable)
-          .where(eq(TelegramConnectionTable.organizationId, organizationId)))
-          .map((row) => row.id)
-        if (telegramConnectionIds.length > 0) {
-          await tx.delete(TelegramPairingTable).where(inArray(TelegramPairingTable.connectionId, telegramConnectionIds))
-          await tx.delete(TelegramChatBindingTable).where(inArray(TelegramChatBindingTable.connectionId, telegramConnectionIds))
-          await tx.delete(TelegramUpdateTable).where(inArray(TelegramUpdateTable.connectionId, telegramConnectionIds))
-        }
-
         const memoryIds = (await tx
           .select({ id: MemoryTable.id })
           .from(MemoryTable)
@@ -509,7 +494,6 @@ export function registerDeleteOrganizationRoutes<T extends { Variables: OrgRoute
         await tx.delete(DesktopPolicyMemberTable).where(eq(DesktopPolicyMemberTable.organizationId, organizationId))
         await tx.delete(DesktopPolicyTable).where(eq(DesktopPolicyTable.organizationId, organizationId))
 
-        await tx.delete(TelegramConnectionTable).where(eq(TelegramConnectionTable.organizationId, organizationId))
         await tx.delete(OrganizationDiagnosticCredentialTable).where(eq(OrganizationDiagnosticCredentialTable.organizationId, organizationId))
         await tx.delete(MemoryTable).where(eq(MemoryTable.org_id, organizationId))
