+++ ee/apps/den-api/src/routes/webhooks/index.ts
@@ -1,10 +1,8 @@
 import type { Env, Hono } from "hono"
 import { registerGithubWebhookRoutes } from "./github.js"
 import { registerStripeWebhookRoutes } from "./stripe.js"
-import { registerTelegramWebhookRoutes } from "./telegram.js"
 
 export function registerWebhookRoutes<T extends Env>(app: Hono<T>) {
   registerGithubWebhookRoutes(app)
   registerStripeWebhookRoutes(app)
-  registerTelegramWebhookRoutes(app)
 }
