+++ ee/apps/den-api/src/server.ts
@@ -7,7 +7,6 @@ import { startScimMaintenanceLoop } from "./scim-maintenance.js"
 import { startCloudIdleStopLoop } from "./workers/cloud-lifecycle.js"
 import { startWorkerProvisioningReconcileLoop } from "./workers/reconciler.js"
 import { startGithubSyncWorker } from "./workers/github-sync.js"
-import { startTelegramUpdateDispatcher } from "./capability-sources/telegram-dispatcher.js"
 import { externalMcpClientRuntimeName } from "./capability-sources/external-mcp-client-runtime.js"
 import { startAutomationSchedulerLoop } from "./automations/scheduler-loop.js"
 
@@ -15,7 +14,6 @@ const stopScimMaintenanceLoop = startScimMaintenanceLoop()
 const stopCloudIdleStopLoop = startCloudIdleStopLoop()
 const stopWorkerProvisioningReconcileLoop = startWorkerProvisioningReconcileLoop()
 const stopGithubSyncWorker = startGithubSyncWorker()
-const stopTelegramUpdateDispatcher = startTelegramUpdateDispatcher()
 const automationScheduler = startAutomationSchedulerLoop()
 
 appLogger.info("external mcp implementation selected", { component: "server", runtime: externalMcpClientRuntimeName })
@@ -79,7 +77,6 @@ async function stopBackgroundLoops() {
     stopCloudIdleStopLoop(),
     stopWorkerProvisioningReconcileLoop(),
     stopGithubSyncWorker(),
-    stopTelegramUpdateDispatcher(),
     automationScheduler.stop(),
   ])
