+++ ee/apps/den-api/test/codemode-tools.test.ts
@@ -124,9 +124,12 @@ test("excludes credential-bound native routes from the Den namespace and manifes
           tags: ["Capability Sources"],
         },
       },
-      "/v1/capabilities/telegram/status": {
+      // Synthetic: every shipped /v1/capabilities/* route is a native provider
+      // today, so this guards the generic rule that only native-provider
+      // prefixes are withheld from tools.den.
+      "/v1/capabilities/other-source/status": {
         get: {
-          operationId: "getV1CapabilitiesTelegramStatus",
+          operationId: "getV1CapabilitiesOtherSourceStatus",
           tags: ["Capability Sources"],
         },
       },
@@ -141,10 +144,10 @@ test("excludes credential-bound native routes from the Den namespace and manifes
 
   expect(built.tools.den?.getCapabilitiesGoogleWorkspaceGmailMessages).toBeUndefined()
   expect(built.tools.den?.getCapabilitiesMicrosoft365CalendarEvents).toBeUndefined()
-  expect(built.tools.den?.getCapabilitiesTelegramStatus).toBeDefined()
+  expect(built.tools.den?.getCapabilitiesOtherSourceStatus).toBeDefined()
   expect(built.tools.den?.getWorkers).toBeDefined()
   const manifestPaths = built.manifest.map((entry) => entry.scriptPath)
-  expect(manifestPaths).toContain("tools.den.getCapabilitiesTelegramStatus")
+  expect(manifestPaths).toContain("tools.den.getCapabilitiesOtherSourceStatus")
   expect(manifestPaths).toContain("tools.den.getWorkers")
   // Absence asserted by scriptPath, not by whole-object equality: extra manifest
   // fields (readOnly/authority) would make an object comparison pass vacuously.
