+++ ee/apps/den-web/app/(den)/dashboard/_components/dashboard-overview-screen.tsx
@@ -13,7 +13,7 @@ import { getMcpConnectionsRoute } from "../../_lib/den-org";
 import { useDenFlow } from "../../_providers/den-flow-provider";
 import { useOrgDashboard } from "../_providers/org-dashboard-provider";
 import { ConnectorQuickAddGrid } from "./connector-quick-add-grid";
-import { useMcpConnectionPresets, useMcpConnections, useTelegramCapabilityStatus } from "./mcp-connections-data";
+import { useMcpConnectionPresets, useMcpConnections } from "./mcp-connections-data";
 import { OrganizationDownloadCard } from "./organization-download-card";
 
 /* ── Types ── */
@@ -88,7 +88,6 @@ export function DashboardOverviewScreen() {
   const { user } = useDenFlow();
   const { data: connections = [] } = useMcpConnections();
   const { data: presets = [] } = useMcpConnectionPresets();
-  const telegramConnected = useTelegramCapabilityStatus(true);
 
   const { data: adoption } = useQuery({
     queryKey: ["telemetry", "adoption"],
@@ -135,7 +134,6 @@ export function DashboardOverviewScreen() {
         <ConnectorQuickAddGrid
           connections={connections}
           presets={presets}
-          telegramConnected={telegramConnected.data === true}
           filter=""
           onSelect={(id) => {
             router.push(`${getMcpConnectionsRoute(activeOrg?.slug)}?quickAdd=${encodeURIComponent(id)}`);
