+++ ee/packages/den-db/src/schema/index.ts
@@ -14,7 +14,6 @@ export * from "./sharables/llm-providers"
 export * from "./sharables/plugin-arch"
 export * from "./subscriptions"
 export * from "./teams"
-export * from "./telegram"
 export * from "./temp-files"
 export * from "./workers"
 export * from "./system"
