+++ ee/packages/utils/src/typeid.ts
@@ -94,10 +94,6 @@ export const idTypesMapNameToPrefix = {
   externalMcpConnection: "emc",
   externalMcpConnectionAccessGrant: "emg",
   pluginMcpRequirementBinding: "pmr",
-  telegramConnection: "tgc",
-  telegramPairing: "tgp",
-  telegramChatBinding: "tgb",
-  telegramUpdate: "tgu",
   memory: "mem",
   memctx: "mctx",
   tempFile: "tmpf",
