+++ evals/flows/den-dashboard-polish.flow.mjs
@@ -226,7 +226,6 @@ export default {
           assert: async () => {
             await ctx.expectText("Google Workspace");
             await ctx.expectText("Microsoft 365");
-            await ctx.expectText("Telegram");
             const extensionsHref = await ctx.eval(`(() => {
               const link = [...document.querySelectorAll('nav a')]
                 .find((entry) => (entry.textContent ?? '').trim().startsWith('Extensions'));
@@ -237,7 +236,7 @@ export default {
           screenshot: screenshot(
             "dashboard-quick-add",
             "The admin dashboard visibly offers reusable connector Quick add choices.",
-            ["Dashboard", "Quick add", "Google Workspace", "Microsoft 365", "Telegram"],
+            ["Dashboard", "Quick add", "Google Workspace", "Microsoft 365"],
           ),
         });
       },
